backend-server/requirements.txt
bson==0.5.8
jsonrpclib-pelix==0.4.0
pkg-resources==0.0.0
python-dateutil==2.8.0
six==1.12.0
